refactor: update AstBuilderNode to use a specific root ast node (#128)
