--- packages/app-builder/src/components/AstBuilder/Root/index.ts
@@ -1,2 +1,0 @@
-export * from './RootAnd';
-export * from './RootOrWithAnd';
