+++ packages/app-builder/src/components/AstBuilder/RootAstBuilderNode/index.ts
@@ -1,0 +1,1 @@
+export * from './RootAstBuilderNode';
